refactor assets repeater into smaller parts
