+++ libs/application/templates/inheritance-report/src/fields/EstateAndVehiclesRepeater/index.tsx
@@ -1,23 +1,7 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import { FC, useState, useEffect, useCallback } from 'react'
-import { useFieldArray, useFormContext, useWatch } from 'react-hook-form'
-import { useLazyQuery } from '@apollo/client'
-import { GET_VEHICLE_QUERY, SEARCH_FOR_PROPERTY_QUERY } from '../../graphql'
-import {
-  GetVehicleInput,
-  Query,
-  SearchForPropertyInput,
-} from '@island.is/api/schema'
-import { InputController } from '@island.is/shared/form-fields'
+import { useFieldArray, useFormContext } from 'react-hook-form'
 import { FieldBaseProps } from '@island.is/application/types'
-import {
-  Box,
-  GridColumn,
-  GridRow,
-  Button,
-  Input,
-  Text,
-} from '@island.is/island-ui/core'
+import { Box, GridRow, Button, Input, Text } from '@island.is/island-ui/core'
 import { Answers } from '../../types'
 import { getErrorViaPath, getValueViaPath } from '@island.is/application/core'
 import { formatCurrency } from '@island.is/application/ui-components'
@@ -26,30 +10,15 @@ import { m } from '../../lib/messages'
 import DoubleColumnRow from '../../components/DoubleColumnRow'
 import {
   getEstateDataFromApplication,
-  isValidRealEstate,
   valueToNumber,
   getDeceasedWasMarriedAndHadAssets,
-  parseLabel,
 } from '../../lib/utils/helpers'
 import { InheritanceReportAsset } from '@island.is/clients/syslumenn'
-import ShareInput from '../../components/ShareInput'
 import DeceasedShare from '../../components/DeceasedShare'
 import { PREPAID_INHERITANCE } from '../../lib/constants'
 import * as styles from '../styles.css'
-
-type RepeaterProps = {
-  field: {
-    props: {
-      sectionTitle?: string
-      sectionTitleVariant?: string
-      fields: Array<object>
-      repeaterButtonText: string
-      sumField: string
-      calcWithShareValue?: boolean
-      assetKey?: string
-    }
-  }
-}
+import { FieldComponent } from './FieldComponent'
+import { RepeaterProps } from './types'
 
 export const EstateAndVehiclesRepeater: FC<
   React.PropsWithChildren<FieldBaseProps<Answers> & RepeaterProps>
@@ -293,286 +262,3 @@ export const EstateAndVehiclesRepeater: FC<
 }
 
 export default EstateAndVehiclesRepeater
-
-interface FieldComponentProps {
-  assetKey?: string
-  onAfterChange?: () => void
-  setLoadingFieldName?: (fieldName: string | null) => void
-  loadingFieldName?: string | null
-  pushRight?: boolean
-  field: Record<string, any>
-  fieldIndex: string
-  fieldName: string
-  error?: string
-  answers?: Answers
-  readOnly?: boolean
-  disabled?: boolean
-}
-
-const FieldComponent = ({
-  assetKey,
-  onAfterChange,
-  setLoadingFieldName,
-  loadingFieldName,
-  pushRight,
-  field,
-  fieldIndex,
-  fieldName,
-  error,
-  answers,
-  readOnly,
-  disabled,
-}: FieldComponentProps) => {
-  const { formatMessage } = useLocale()
-
-  let content = null
-
-  const defaultProps = {
-    ...field,
-    id: fieldName,
-    name: fieldName,
-    format: field.format,
-    label: formatMessage(parseLabel(field.title, answers)),
-    defaultValue: '',
-    type: field.type,
-    placeholder: field.placeholder,
-    backgroundColor: field.color ? field.color : 'blue',
-    currency: field.currency,
-    required: readOnly ? false : field.required,
-    loading: fieldName === loadingFieldName,
-    suffix: field.suffix,
-    onChange: () => onAfterChange?.(),
-    error: error,
-    readOnly: readOnly,
-    disabled: disabled,
-  }
-
-  switch (field.id) {
-    case 'sectionTitle':
-      return (
-        <GridColumn key={fieldName} span="1/1">
-          <Text
-            variant={
-              field.sectionTitleVariant ? field.sectionTitleVariant : 'h5'
-            }
-            marginBottom={2}
-          >
-            {field.sectionTitle}
-          </Text>
-        </GridColumn>
-      )
-
-    case 'assetNumber':
-      if (assetKey === 'assets') {
-        content = (
-          <RealEstateNumberField
-            field={field}
-            fieldName={fieldName}
-            fieldIndex={fieldIndex}
-            setLoadingFieldName={setLoadingFieldName}
-            {...defaultProps}
-          />
-        )
-      } else if (assetKey === 'vehicles') {
-        content = (
-          <VehicleNumberField
-            field={field}
-            fieldName={fieldName}
-            fieldIndex={fieldIndex}
-            setLoadingFieldName={setLoadingFieldName}
-            {...defaultProps}
-          />
-        )
-      }
-      break
-
-    case 'share':
-      content = (
-        <ShareInput
-          name={fieldName}
-          label={formatMessage(m.propertyShare)}
-          onAfterChange={onAfterChange}
-          readOnly={readOnly}
-          hasError={!!error}
-          required={field.required && !readOnly}
-          disabled={disabled}
-        />
-      )
-      break
-
-    default:
-      content = <InputController {...defaultProps} />
-
-      break
-  }
-
-  return (
-    <DoubleColumnRow
-      span={field.width === 'full' ? ['1/1', '1/1'] : ['1/1', '1/2']}
-      pushRight={pushRight}
-      paddingBottom={2}
-      key={fieldName}
-    >
-      {content}
-    </DoubleColumnRow>
-  )
-}
-
-const RealEstateNumberField = ({
-  fieldIndex,
-  fieldName,
-  error,
-  setLoadingFieldName,
-  ...props
-}: FieldComponentProps) => {
-  const { formatMessage } = useLocale()
-  const { setValue, clearErrors, setError } = useFormContext()
-
-  const descriptionFieldName = `${fieldIndex}.description`
-
-  const propertyNumberInput = useWatch({
-    name: fieldName,
-    defaultValue: '',
-  })
-
-  const [getProperty, { loading: queryLoading, error: _queryError }] =
-    useLazyQuery<Query, { input: SearchForPropertyInput }>(
-      SEARCH_FOR_PROPERTY_QUERY,
-      {
-        onCompleted: (data) => {
-          if (isValidRealEstate(propertyNumberInput)) {
-            clearErrors(descriptionFieldName)
-            setValue(
-              descriptionFieldName,
-              data.searchForProperty?.defaultAddress?.display ?? '',
-            )
-          } else {
-            setError(fieldName, {
-              message: formatMessage(m.errorPropertyNumber),
-              type: 'validate',
-            })
-          }
-        },
-        fetchPolicy: 'network-only',
-      },
-    )
-
-  useEffect(() => {
-    setLoadingFieldName?.(queryLoading ? descriptionFieldName : null)
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [queryLoading])
-
-  useEffect(() => {
-    if (!props.readOnly) {
-      const propertyNumber = propertyNumberInput
-        .trim()
-        .toUpperCase()
-        .replace('-', '')
-
-      setValue(descriptionFieldName, '')
-
-      if (isValidRealEstate(propertyNumber)) {
-        clearErrors(fieldName)
-
-        getProperty({
-          variables: {
-            input: {
-              propertyNumber,
-            },
-          },
-        })
-      } else {
-        if (propertyNumber.length !== 0) {
-          setError(fieldName, {
-            message: formatMessage(m.errorPropertyNumber),
-            type: 'validate',
-          })
-        }
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [propertyNumberInput])
-
-  return (
-    <InputController
-      id={fieldName}
-      name={fieldName}
-      label={formatMessage(m.propertyNumber)}
-      defaultValue={propertyNumberInput}
-      error={error ? formatMessage(m.errorPropertyNumber) : undefined}
-      placeholder={propertyNumberInput > 0 ? '' : props.field.placeholder}
-      {...props}
-    />
-  )
-}
-
-const VehicleNumberField = ({
-  fieldIndex,
-  fieldName,
-  error,
-  setLoadingFieldName,
-  ...props
-}: FieldComponentProps) => {
-  const { formatMessage } = useLocale()
-  const { setValue, clearErrors } = useFormContext()
-
-  const descriptionFieldName = `${fieldIndex}.description`
-
-  const assetNumberInput = useWatch({
-    name: fieldName,
-    defaultValue: '',
-  })
-
-  const [getProperty, { loading: queryLoading, error: _queryError }] =
-    useLazyQuery<Query, { input: GetVehicleInput }>(GET_VEHICLE_QUERY, {
-      onError: (_e) => {
-        setValue(descriptionFieldName, '')
-      },
-      onCompleted: (data) => {
-        const carName =
-          `${data?.syslumennGetVehicle?.manufacturer} ${data.syslumennGetVehicle?.modelName}`.trim()
-        if (
-          carName.length === 0 ||
-          carName.startsWith('null') ||
-          carName.endsWith('null')
-        ) {
-          setValue(descriptionFieldName, '')
-          return
-        }
-        clearErrors(descriptionFieldName)
-        setValue(descriptionFieldName, carName)
-      },
-      fetchPolicy: 'network-only',
-    })
-
-  useEffect(() => {
-    setLoadingFieldName?.(queryLoading ? descriptionFieldName : null)
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [queryLoading])
-
-  useEffect(() => {
-    if (assetNumberInput.trim().length > 0) {
-      getProperty({
-        variables: {
-          input: {
-            vehicleId: assetNumberInput.trim().toUpperCase(),
-          },
-        },
-      })
-    } else {
-      setValue(descriptionFieldName, '')
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [assetNumberInput])
-
-  return (
-    <InputController
-      id={fieldName}
-      name={fieldName}
-      label={formatMessage(m.propertyNumber)}
-      defaultValue={assetNumberInput}
-      error={error}
-      {...props}
-    />
-  )
-}
