+++ libs/application/templates/inheritance-report/src/fields/EstateAndVehiclesRepeater/types.ts
@@ -1,0 +1,30 @@
+import { Answers } from '../../types'
+
+export type RepeaterProps = {
+  field: {
+    props: {
+      sectionTitle?: string
+      sectionTitleVariant?: string
+      fields: Array<object>
+      repeaterButtonText: string
+      sumField: string
+      calcWithShareValue?: boolean
+      assetKey?: string
+    }
+  }
+}
+
+export interface FieldComponentProps {
+  assetKey?: string
+  onAfterChange?: () => void
+  setLoadingFieldName?: (fieldName: string | null) => void
+  loadingFieldName?: string | null
+  pushRight?: boolean
+  field: Record<string, any>
+  fieldIndex: string
+  fieldName: string
+  error?: string
+  answers?: Answers
+  readOnly?: boolean
+  disabled?: boolean
+}
